refactor(contact-list): split files over the 500-line soft cap
Per CodeRabbit's outside-diff review + the new overage my negative-logging
change introduced. No behavior change; pure module moves with re-exports so
no external import path changes.

- handlers/crud.rs (507) -> handlers/header_ops.rs (create/delete/rename/
  flags) + handlers/member_ops.rs (add/remove); mod.rs re-exports.
- base/contact_list/persistence.rs (596) -> persistence/mod.rs + tests.rs
  (production/tests seam, matching the helpers/ precedent).
- cell/cell_methods/contact_list.rs (519, pushed over by the send-failure
  logging) -> contact_list/mod.rs + tests.rs (same prod/tests seam).
